integrated_rag_clean/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.3
openai==1.37.1
langchain==0.2.16
langchain-community==0.2.16
faiss-cpu==1.8.0.post1
pypdf==4.3.1
moviepy==1.0.3
openai-whisper==20231117
requests==2.32.3
beautifulsoup4==4.12.3
